moves package to models library
